Move user stuff into own folder
